Directly embed a Read+Seek object into the BTree struct
This was to help simplify the code from needing to clone and constantly
borrow the object and to also enforce the idea that the BTree now owns
it. Nothing else needs to (or should) access the underlying file-like
object once the B-Tree gets it. Also, it no longer requires it to be a
Fork so it's easier to mock in unit tests.

However, this now requires that B-Trees be mutable even on Read
operations as Read operations implies updating a position inside the
handler. This has required a lot more code changes than expected and
made it more awkward in some ways to access. This is probably for the
better in the long run as caching will probably be implemented requiring
mutation and eventually Write access.